project-charter-generator/backend/requirements.txt
python-dotenv
openai
langchain
langchain-community
sentence-transformers
faiss-cpu
streamlit
Flask
flask-cors
httpx
PyJWT
requests
pytest
